foundations/projects/project-1.2-prompt-playground/requirements.txt
langchain-core>=0.3.0
langchain-openai>=0.3.0
pydantic>=2.7.0
python-dotenv>=1.0.1
